+++ changelog.md
@@ -24,6 +24,7 @@
 * LLT-4509: Enable windows arm64 build.
 * LLT-5007: Bump moose tracker to v3.0.0-libtelioApp
 * LLT-4873: Remove old SWIG FFI bindings
+* LLT-4996: nurse: Stop adding virtual peers to qos
 
 <br>